Huge xmonad overhaul to integrate solarized theme
This is a big change to the xmonad/xmobar configuration that, among
other things, integrates more solarized colors into visual elements of
the program. Other changes include using dmenu as the program launcher,
adding more monitors to xmobar, and adding an icon to xmobar that will
switch to a random wallpaper when clicked.